requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
coloredlogs==15.0.1
frozenlist==1.4.1
humanfriendly==10.0
idna==3.7
multidict==6.0.5
py-cord==2.6.0
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.9.4
